+++ internal/bot/components/misc/misc.go
@@ -153,7 +153,7 @@ func Init(kit *route.Kit, runState *state.State) error {
 		Name:        "Restart bot",
 		CommandText: []string{"restart"},
 		Restrictions: []route.CommandRestriction{func(session *discordgo.Session, message *discordgo.MessageCreate) (bool, error) {
-			return message.Author.ID == config.OwnerId, nil
+			return common.IsOwner(message.Author.ID), nil
 		}},
 		Run: func(ctx *route.MessageContext) error {
 			log.Info().Msgf("Restarting request of %s %s", ctx.Message.Author.ID, ctx.Message.Author.String())
@@ -176,7 +176,7 @@ func Init(kit *route.Kit, runState *state.State) error {
 		Name:        "Goroutine stacktrace",
 		CommandText: []string{"goroutinestack"},
 		Restrictions: []route.CommandRestriction{func(session *discordgo.Session, message *discordgo.MessageCreate) (bool, error) {
-			return message.Author.ID == config.OwnerId, nil
+			return common.IsOwner(message.Author.ID), nil
 		}},
 		Run:       comp.GoroutineStack,
 		Invisible: true,